refactor(openai): split provider module into submodules
