+++ model_gateway/src/routers/openai/provider/gemini.rs
@@ -1,0 +1,29 @@
+use serde_json::Value;
+
+use super::{Provider, ProviderError};
+use crate::worker::{Endpoint, ProviderType};
+
+pub struct GeminiProvider;
+
+impl Provider for GeminiProvider {
+    fn provider_type(&self) -> ProviderType {
+        ProviderType::Gemini
+    }
+
+    fn transform_request(
+        &self,
+        payload: &mut Value,
+        endpoint: Endpoint,
+    ) -> Result<(), ProviderError> {
+        super::types::strip_sglang_fields(payload);
+
+        if endpoint == Endpoint::Chat {
+            if let Some(obj) = payload.as_object_mut() {
+                if obj.get("logprobs").and_then(|v| v.as_bool()) == Some(false) {
+                    obj.remove("logprobs");
+                }
+            }
+        }
+        Ok(())
+    }
+}
