+++ model_gateway/src/routers/openai/provider/mod.rs
@@ -1,0 +1,19 @@
+//! Provider abstractions for vendor-specific API transformations.
+
+mod anthropic;
+mod gemini;
+mod openai;
+mod provider_trait;
+mod registry;
+mod sglang;
+mod types;
+mod xai;
+
+pub use anthropic::AnthropicProvider;
+pub use gemini::GeminiProvider;
+pub use openai::OpenAIProvider;
+pub use provider_trait::Provider;
+pub use registry::ProviderRegistry;
+pub use sglang::SGLangProvider;
+pub use types::ProviderError;
+pub use xai::XAIProvider;
