+++ model_gateway/src/routers/openai/provider/openai.rs
@@ -1,0 +1,10 @@
+use super::Provider;
+use crate::worker::ProviderType;
+
+pub struct OpenAIProvider;
+
+impl Provider for OpenAIProvider {
+    fn provider_type(&self) -> ProviderType {
+        ProviderType::OpenAI
+    }
+}
